Enable running of dp without SSH setup
If the user does not have access to dp-setup the cli would previously
fail, blocking use of any of the other features of the tool. By changing
this error to warning we enable a wider range of users who may not have
the dp-setup repo cloned to still use the other subcommands.