Allow repo paths to be customised
Allow the user to customise where the dp-hierarchy-api and
dp-code-list-scripts repos are checked out. This is mostly to support
the fact that these repos will need to move outside the GOPATH in order
to work for modules. Prior to this change these repos would have to be
checked out twice.