Fix ansible groups parsing after ansible upgrade
During the upgrade to ansible 2.10 the groups in the `hosts` file were
changed. Long term it might be best to change the way we are getting the
groups or even included them within this repo, but in the meantime the
parsing has been updated to work with the latest ansible inventory.